# docs/openapi-identity.yaml
openapi: 3.1.0
info:
  title: Pulso Piura API
  version: 0.8.0
servers:
  - url: http://localhost:8080/api/v1
paths:
...
  /venues:
    get:
      summary: Busca sedes publicadas con disponibilidad teórica
      parameters:
        - {in: query, name: district, schema: {type: string}}
        - {in: query, name: sport, schema: {type: string}}
        - {in: query, name: date, schema: {type: string, format: date}}
        - {in: query, name: page, schema: {type: integer, minimum: 0, default: 0}}
        - {in: query, name: size, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
      responses:
        "200":
          description: Página de sedes publicadas; si se indica fecha, conserva solo sedes con alguna franja teórica libre
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PublicVenuePage"}
        "400": {$ref: "#/components/responses/BadRequest"}
  /venues/{publicSlug}:
    parameters:
      - {in: path, name: publicSlug, required: true, schema: {type: string}}
    get:
      summary: Obtiene una sede publicada por su identificador público global
      responses:
        "200": {description: Sede publicada, content: {application/json: {schema: {$ref: "#/components/schemas/PublicVenue"}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /venues/{publicSlug}/spaces:
    parameters:
      - {in: path, name: publicSlug, required: true, schema: {type: string}}
    get:
      summary: Lista las canchas publicadas de una sede publicada
      responses:
        "200": {description: Canchas públicas, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/PublicSportSpace"}}}}}
        "404": {$ref: "#/components/responses/NotFound"}
  /spaces/{spaceId}/availability:
    parameters:
      - $ref: "#/components/parameters/SpaceId"
      - {in: query, name: date, required: true, schema: {type: string, format: date}}
    get:
      summary: Calcula franjas teóricas disponibles para una cancha publicada
      description: Aplica reglas semanales, cierres, mantenimiento y precios especiales en la zona horaria de la organización. Aún no descuenta reservas.
      responses:
        "200": {description: Disponibilidad teórica, content: {application/json: {schema: {$ref: "#/components/schemas/PublicAvailability"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "404": {$ref: "#/components/responses/NotFound"}
...
  /organizations/{organizationId}/venues:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
    get:
      summary: Lista las sedes internas de una organización autorizada
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Sedes del tenant
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/Venue"}}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      summary: Crea una sede en estado borrador
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VenueInput"}
      responses:
        "201":
          description: Sede creada
          content:
            application/json:
              schema: {$ref: "#/components/schemas/Venue"}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /organizations/{organizationId}/venues/{venueId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/VenueId"
    put:
      summary: Edita una sede mediante control optimista de versión
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/VenueUpdateInput"}
      responses:
        "200": {description: Sede actualizada, content: {application/json: {schema: {$ref: "#/components/schemas/Venue"}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
    delete:
      summary: Archiva una sede sin eliminar su historial
      security: [{ bearerAuth: [] }]
      responses:
        "204": {description: Sede archivada}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
  /organizations/{organizationId}/spaces/{spaceId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/SpaceId"
    put:
      summary: Edita una cancha mediante control optimista de versión
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/SportSpaceUpdateInput"}
      responses:
        "200": {description: Cancha actualizada, content: {application/json: {schema: {$ref: "#/components/schemas/SportSpace"}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
    delete:
      summary: Archiva una cancha
      security: [{ bearerAuth: [] }]
      responses:
        "204": {description: Cancha archivada}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /organizations/{organizationId}/spaces/{spaceId}/publish:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/SpaceId"
    post:
      summary: Publica una cancha cuando su sede ya está publicada
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Cancha publicada, content: {application/json: {schema: {$ref: "#/components/schemas/SportSpace"}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /organizations/{organizationId}/spaces/{spaceId}/availability-rules:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/SpaceId"
    get:
      summary: Lista reglas semanales de la cancha
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Reglas configuradas, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/AvailabilityRule"}}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      summary: Crea una regla semanal sin superponerse con reglas activas
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AvailabilityRuleInput"}
      responses:
        "201": {description: Regla creada, content: {application/json: {schema: {$ref: "#/components/schemas/AvailabilityRule"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {$ref: "#/components/responses/Forbidden"}
        "409": {$ref: "#/components/responses/Conflict"}
  /organizations/{organizationId}/spaces/{spaceId}/availability-rules/{ruleId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/SpaceId"
      - {in: path, name: ruleId, required: true, schema: {type: string, format: uuid}}
      - $ref: "#/components/parameters/Version"
    delete:
      summary: Desactiva una regla semanal
      security: [{ bearerAuth: [] }]
      responses:
        "204": {description: Regla desactivada}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
  /organizations/{organizationId}/spaces/{spaceId}/exceptions:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/SpaceId"
    get:
      summary: Lista excepciones puntuales de la cancha
      security: [{ bearerAuth: [] }]
      responses:
        "200": {description: Excepciones configuradas, content: {application/json: {schema: {type: array, items: {$ref: "#/components/schemas/AvailabilityException"}}}}}
        "403": {$ref: "#/components/responses/Forbidden"}
    post:
      summary: Crea un cierre, mantenimiento o precio especial
      description: OPERATOR puede crear cierres y mantenimiento; precios especiales requieren OWNER o ADMIN.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/AvailabilityExceptionInput"}
      responses:
        "201": {description: Excepción creada, content: {application/json: {schema: {$ref: "#/components/schemas/AvailabilityException"}}}}
        "400": {$ref: "#/components/responses/BadRequest"}
        "403": {$ref: "#/components/responses/Forbidden"}
  /organizations/{organizationId}/spaces/{spaceId}/exceptions/{exceptionId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/SpaceId"
      - {in: path, name: exceptionId, required: true, schema: {type: string, format: uuid}}
      - $ref: "#/components/parameters/Version"
    delete:
      summary: Cancela lógicamente una excepción
      security: [{ bearerAuth: [] }]
      responses:
        "204": {description: Excepción cancelada}
        "403": {$ref: "#/components/responses/Forbidden"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationId: {in: path, name: organizationId, required: true, schema: {type: string, format: uuid}}
    VenueId: {in: path, name: venueId, required: true, schema: {type: string, format: uuid}}
    SpaceId: {in: path, name: spaceId, required: true, schema: {type: string, format: uuid}}
    Version: {in: query, name: version, required: true, schema: {type: integer, format: int64, minimum: 0}}
  responses:
    BadRequest: {description: Entrada inválida, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    Forbidden: {description: Sin membresía o permiso en la organización, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    NotFound: {description: Recurso no encontrado dentro del tenant, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
    Conflict: {description: Estado, versión o dato en conflicto, content: {application/problem+json: {schema: {$ref: "#/components/schemas/Problem"}}}}
  schemas:
    Problem:
      type: object
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
    VenueInput:
      type: object
      required: [name, address, districtCode]
      properties:
        name: {type: string, maxLength: 160}
        address: {type: string, maxLength: 240}
        districtCode: {type: string, maxLength: 60}
        latitude: {type: number, minimum: -90, maximum: 90}
        longitude: {type: number, minimum: -180, maximum: 180}
        publicPhone: {type: string, maxLength: 30}
        amenityCodes:
          type: array
          uniqueItems: true
          items: {type: string, maxLength: 40}
    VenueUpdateInput:
      allOf:
        - $ref: "#/components/schemas/VenueInput"
        - type: object
          required: [version]
          properties:
            version: {type: integer, format: int64, minimum: 0}
    Venue:
      allOf:
        - $ref: "#/components/schemas/VenueUpdateInput"
        - type: object
          required: [id, organizationId, slug, publicSlug, status]
          properties:
            id: {type: string, format: uuid}
            organizationId: {type: string, format: uuid}
            slug: {type: string}
            publicSlug: {type: string, description: Identificador global estable para rutas públicas}
            status: {type: string, enum: [DRAFT, PUBLISHED, ARCHIVED]}
    SportSpaceInput:
      type: object
      required: [name, sportCode, formatCode, capacity, indoor]
      properties:
        name: {type: string, maxLength: 160}
        sportCode: {type: string, enum: [FOOTBALL, VOLLEYBALL, BASKETBALL, PADEL, TENNIS]}
        formatCode: {type: string, maxLength: 40}
        capacity: {type: integer, minimum: 1}
        surfaceType: {type: string, maxLength: 40}
        indoor: {type: boolean}
        amenityCodes:
          type: array
          uniqueItems: true
          items: {type: string, maxLength: 40}
    SportSpaceUpdateInput:
      allOf:
        - $ref: "#/components/schemas/SportSpaceInput"
        - type: object
          required: [version]
          properties:
            version: {type: integer, format: int64, minimum: 0}
    SportSpace:
      allOf:
        - $ref: "#/components/schemas/SportSpaceUpdateInput"
        - type: object
          required: [id, organizationId, venueId, status]
          properties:
            id: {type: string, format: uuid}
            organizationId: {type: string, format: uuid}
            venueId: {type: string, format: uuid}
            status: {type: string, enum: [DRAFT, PUBLISHED, ARCHIVED]}
...
    PublicVenuePage:
      type: object
      required: [items, page, size, total]
      properties:
        items: {type: array, items: {$ref: "#/components/schemas/PublicVenue"}}
        page: {type: integer, minimum: 0}
        size: {type: integer, minimum: 1, maximum: 50}
        total: {type: integer, format: int64, minimum: 0}
    PublicVenue:
      type: object
      required: [publicSlug, name, address, districtCode, amenityCodes]
      properties:
        publicSlug: {type: string}
        name: {type: string}
        address: {type: string}
        districtCode: {type: string}
        latitude: {type: [number, "null"]}
        longitude: {type: [number, "null"]}
        publicPhone: {type: [string, "null"]}
        amenityCodes: {type: array, uniqueItems: true, items: {type: string}}
    PublicSportSpace:
      type: object
      required: [id, name, sportCode, formatCode, capacity, indoor, amenityCodes]
      properties:
        id: {type: string, format: uuid}
        name: {type: string}
        sportCode: {type: string}
        formatCode: {type: string}
        capacity: {type: integer, minimum: 1}
        surfaceType: {type: [string, "null"]}
        indoor: {type: boolean}
        amenityCodes: {type: array, uniqueItems: true, items: {type: string}}
    PublicAvailability:
      type: object
      required: [sportSpaceId, timezone, date, slots]
      properties:
        sportSpaceId: {type: string, format: uuid}
        timezone: {type: string, example: America/Lima}
        date: {type: string, format: date}
        slots: {type: array, items: {$ref: "#/components/schemas/PublicSlot"}}
    PublicSlot:
      type: object
      required: [startsAt, endsAt, priceMinor, currency]
      properties:
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time}
        priceMinor: {type: integer, format: int64, minimum: 0}
        currency: {type: string, const: PEN}
...
    AvailabilityRuleInput:
      type: object
      required: [dayOfWeek, startLocalTime, endLocalTime, slotMinutes, priceMinor, validFrom]
      properties:
        dayOfWeek: {type: integer, minimum: 1, maximum: 7}
        startLocalTime: {type: string, format: time}
        endLocalTime: {type: string, format: time}
        slotMinutes: {type: integer, minimum: 30, maximum: 180}
        priceMinor: {type: integer, format: int64, minimum: 0}
        validFrom: {type: string, format: date}
        validTo: {type: [string, "null"], format: date}
    AvailabilityRule:
      allOf:
        - $ref: "#/components/schemas/AvailabilityRuleInput"
        - type: object
          required: [id, sportSpaceId, currency, status, version]
          properties:
            id: {type: string, format: uuid}
            sportSpaceId: {type: string, format: uuid}
            currency: {type: string, const: PEN}
            status: {type: string, enum: [ACTIVE, INACTIVE]}
            version: {type: integer, format: int64, minimum: 0}
    AvailabilityExceptionInput:
      type: object
      required: [startsAt, endsAt, type]
      properties:
        startsAt: {type: string, format: date-time}
        endsAt: {type: string, format: date-time}
        type: {type: string, enum: [CLOSED, MAINTENANCE, SPECIAL_PRICE]}
        priceMinor: {type: [integer, "null"], format: int64, minimum: 0}
        reason: {type: [string, "null"], maxLength: 240}
    AvailabilityException:
      allOf:
        - $ref: "#/components/schemas/AvailabilityExceptionInput"
        - type: object
          required: [id, sportSpaceId, status, version]
          properties:
            id: {type: string, format: uuid}
            sportSpaceId: {type: string, format: uuid}
            currency: {type: [string, "null"], enum: [PEN, null]}
            status: {type: string, enum: [ACTIVE, CANCELLED]}
            version: {type: integer, format: int64, minimum: 0}